Adds OrdersContext and useOrderContext hook
