+++ frontend/src/App.tsx
@@ -17,10 +17,11 @@ import ProfileUpdate from './components/ProfileUpdate';
 import Footer from './components/footer/Footer';
 
 import 'react-toastify/dist/ReactToastify.css';
-import { AppContextProvider } from './contexts/app-context/AppContext';
+import { AppContextProvider } from './contexts/app-context';
 import ShoppingCartContextProvider from './contexts/shopping-cart-context';
 import UserAuthContextProvider from './contexts/user-auth-context';
 import CheckoutContextProvider from './contexts/checkout-context';
+import OrdersContextProvider from './contexts/orders-context';
 
 const reactQueryClient = new QueryClient()
 
@@ -31,26 +32,28 @@ function App() {
         <ShoppingCartContextProvider>
           <CheckoutContextProvider>
             <UserAuthContextProvider>
-              <QueryClientProvider client={ reactQueryClient }>
-                <BrowserRouter>
-                  <ToastContainer position="bottom-center" limit={ 1 } />
-                  <PrimarySearchAppBar />
-                  <Routes>
-                    <Route path="/" element={ <DisplayProducts /> } />
-                    <Route path="/product/label/:label" element={ <DisplayProduct /> } />
-                    <Route path="/cart" element={ <CartPage /> } />
-                    <Route path="/user/signin" element={ <SignIn /> } />
-                    <Route path="/user/signup" element={ <SignUp /> } />
-                    <Route path="/user/update" element={ <ProfileUpdate /> } />
-                    <Route path="/shipping" element={ <ShippingAddressUi /> } />
-                    <Route path="/payment" element={ <PaymentUi /> } />
-                    <Route path="/placeOrder" element={ <PlaceOrderUi /> } />
-                    <Route path="/order/:id" element={ <OrderScreen /> } />
-                    <Route path="/ordershistory" element={ <OrdersHistory /> } />
-                  </Routes>
-                  <Footer />
-                </BrowserRouter>
-              </QueryClientProvider>
+              <OrdersContextProvider>
+                <QueryClientProvider client={ reactQueryClient }>
+                  <BrowserRouter>
+                    <ToastContainer position="bottom-center" limit={ 1 } />
+                    <PrimarySearchAppBar />
+                    <Routes>
+                      <Route path="/" element={ <DisplayProducts /> } />
+                      <Route path="/product/label/:label" element={ <DisplayProduct /> } />
+                      <Route path="/cart" element={ <CartPage /> } />
+                      <Route path="/user/signin" element={ <SignIn /> } />
+                      <Route path="/user/signup" element={ <SignUp /> } />
+                      <Route path="/user/update" element={ <ProfileUpdate /> } />
+                      <Route path="/shipping" element={ <ShippingAddressUi /> } />
+                      <Route path="/payment" element={ <PaymentUi /> } />
+                      <Route path="/placeOrder" element={ <PlaceOrderUi /> } />
+                      <Route path="/order/:id" element={ <OrderScreen /> } />
+                      <Route path="/ordershistory" element={ <OrdersHistory /> } />
+                    </Routes>
+                    <Footer />
+                  </BrowserRouter>
+                </QueryClientProvider>
+              </OrdersContextProvider>
             </UserAuthContextProvider>
           </CheckoutContextProvider>
         </ShoppingCartContextProvider>
