+++ frontend/src/components/checkout/place-order/OrderSummary.tsx
@@ -1,3 +1,5 @@
+// requires refactoring 
+
 import { Box, Button, CardContent, Divider, Typography } from "@mui/material";
 
 import Card from "@mui/material/Card";
