+++ frontend/src/components/checkout/styles.tsx
@@ -1,3 +1,5 @@
+
+// requires refactoring 
 export const shoppingAddressStyles = {
     mainContainer: {
         display: 'flex',
